Split src/resources/networks into directory module
... separating resolveRelations and resolvePossibleRelations

- For each, precompute and export ancestor/descendant versions

- Extract ancestor vs. descendant specific options into respective
  `relationshipProperties` functions

- Move logging to the core `resolve` functions to make it so that
  DEBUG=db:resources:networks gives high-levels and not internals
